Biases, multiple activations, and get/set params.
Made the following changes to the branch.
Added Biases to the NeuralNet model, including to the forward and the backprop parts.
Fixed the missing initialization bug for having multiple different activation functions for each later.
Added a setParams() and getParams() to the NeuralNet class (which you don’t use in tests, but seems useful enough).
I tested them with your backprop xor problem, and it works. It works very well with the tanh. It works semi-okay with the other activation functions.